Move agent registry into runtime package
