+++ agent/helpudoc_agent/runtime/__init__.py
@@ -1,0 +1,3 @@
+"""Agent runtime subpackage (registry, future state/memory moves)."""
+from __future__ import annotations
+
